rate limiter

+++ src/main/java/demo/security/servlet/FileServlet.java
@@ -1,5 +1,6 @@
 package demo.security.servlet;
 
+import demo.security.util.RateLimitFilter;
 import demo.security.util.Utils;
 
 import javax.servlet.ServletException;
@@ -13,6 +14,7 @@ import java.io.IOException;
 public class FileServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        if (!RateLimitFilter.allowRequest(request, response)) return;
         String data = request.getParameter("data");
         Utils.deleteFile(data);
     }
